fix(ui): refactor for combined near full screen for debug panel
