+++ ui/src/features/Editor/components/OverlayUI/components/DebugPanel/DebugLogs/hooks.ts
@@ -1,0 +1,21 @@
+import { useMemo } from "react";
+
+import { useIndexedDB } from "@flow/lib/indexedDB";
+import { useCurrentProject } from "@flow/stores";
+
+export default () => {
+  const [currentProject] = useCurrentProject();
+
+  const { value: debugRunState } = useIndexedDB("debugRun");
+
+  const debugJobId = useMemo(
+    () =>
+      debugRunState?.jobs?.find((job) => job.projectId === currentProject?.id)
+        ?.jobId,
+    [debugRunState, currentProject],
+  );
+
+  return {
+    debugJobId,
+  };
+};
