+++ ui/src/features/Editor/components/OverlayUI/components/DebugPanel/DebugPreview/hooks.ts
@@ -1,0 +1,15 @@
+import { useState } from "react";
+
+export default () => {
+  const [minimized, setMinimized] = useState(false);
+
+  const handleTabChange = () => {
+    if (minimized) {
+      setMinimized(false);
+    }
+  };
+
+  return {
+    handleTabChange,
+  };
+};
