+++ ui/src/features/Editor/components/OverlayUI/components/index.ts
@@ -3,6 +3,7 @@ export { default as Infobar } from "./Infobar";
 export { default as Toolbox } from "./Toolbox";
 export { default as NodePickerDialog } from "./NodePickerDialog";
 export { default as LayoutOptionsDialog } from "./LayoutOptionsDialog";
-export { default as DebugLogs } from "./DebugLogs";
-export { default as DebugPreview } from "./DebugPreview";
+export { default as DebugLogs } from "./DebugPanel/DebugLogs";
+export { default as DebugPreview } from "./DebugPanel/DebugPreview";
 export { default as JobStatus } from "./JobStatus";
+export { default as DebugPanel } from "./DebugPanel";
