+++ ui/src/lib/i18n/locales/en.json
@@ -72,18 +72,20 @@
   "All nodes in viewport": "",
   "Exit fullscreen": "",
   "Enter fullscreen": "",
-  "Workflow Logs": "",
   "Table Viewer Could Not Be Loaded. Check if the data is valid.": "",
   "3D Viewer Could Not Be Loaded. Check if the data is valid.": "",
   "2D Viewer Could Not Be Loaded. Check if the data is valid.": "",
-  "Output data": "",
-  "Select Data to Preview": "",
-  "Table Viewer": "",
   "2D Viewer": "",
   "3D Viewer": "",
   "Your workflow completed without any output data.": "",
   "Please review the logs to see if there were any errors during the workflow process.": "",
   "OK": "",
+  "Output data": "",
+  "Debug Run": "",
+  "Workflow Logs": "",
+  "Data Preview": "",
+  "Tabular Viewer": "",
+  "Select Data to Preview": "",
   "Edge ID: ": "",
   "Source Node ID: ": "",
   "Target Node ID: ": "",
