+++ ui/src/lib/i18n/locales/es.json
@@ -72,18 +72,20 @@
   "All nodes in viewport": "Todos los nodos en la vista",
   "Exit fullscreen": "Salir de pantalla completa",
   "Enter fullscreen": "Entrar en pantalla completa",
-  "Workflow Logs": "Registros del flujo de trabajo",
   "Table Viewer Could Not Be Loaded. Check if the data is valid.": "",
   "3D Viewer Could Not Be Loaded. Check if the data is valid.": "",
   "2D Viewer Could Not Be Loaded. Check if the data is valid.": "",
-  "Output data": "Datos de salida",
-  "Select Data to Preview": "Seleccionar datos para previsualizar",
-  "Table Viewer": "Visor de tabla",
   "2D Viewer": "",
   "3D Viewer": "Visor 3D",
   "Your workflow completed without any output data.": "",
   "Please review the logs to see if there were any errors during the workflow process.": "",
   "OK": "Aceptar",
+  "Output data": "Datos de salida",
+  "Debug Run": "",
+  "Workflow Logs": "Registros del flujo de trabajo",
+  "Data Preview": "",
+  "Tabular Viewer": "",
+  "Select Data to Preview": "Seleccionar datos para previsualizar",
   "Edge ID: ": "ID de borde: ",
   "Source Node ID: ": "ID del nodo fuente: ",
   "Target Node ID: ": "ID del nodo de destino: ",
