+++ ui/src/lib/i18n/locales/fr.json
@@ -72,18 +72,20 @@
   "All nodes in viewport": "Tous les nœuds dans le viewport",
   "Exit fullscreen": "Quitter le plein écran",
   "Enter fullscreen": "Passer en plein écran",
-  "Workflow Logs": "Journaux du flux de travail",
   "Table Viewer Could Not Be Loaded. Check if the data is valid.": "",
   "3D Viewer Could Not Be Loaded. Check if the data is valid.": "",
   "2D Viewer Could Not Be Loaded. Check if the data is valid.": "",
-  "Output data": "Sortie de données",
-  "Select Data to Preview": "Sélectionner les données à prévisualiser",
-  "Table Viewer": "Tableau de visualisation",
   "2D Viewer": "",
   "3D Viewer": "Visualiseur 3D",
   "Your workflow completed without any output data.": "Votre flux de travail s'est terminé sans aucune donnée de sortie.",
   "Please review the logs to see if there were any errors during the workflow process.": "Veuillez consulter les journaux pour voir s'il y a eu des erreurs pendant le processus de flux de travail.",
   "OK": "OK",
+  "Output data": "Sortie de données",
+  "Debug Run": "",
+  "Workflow Logs": "Journaux du flux de travail",
+  "Data Preview": "",
+  "Tabular Viewer": "",
+  "Select Data to Preview": "Sélectionner les données à prévisualiser",
   "Edge ID: ": "ID de bord: ",
   "Source Node ID: ": "ID du nœud source: ",
   "Target Node ID: ": "ID du nœud de destination: ",
