+++ ui/src/lib/i18n/locales/ja.json
@@ -72,18 +72,20 @@
   "All nodes in viewport": "ビューポート内のすべてのノード",
   "Exit fullscreen": "フルスクリーンを終了",
   "Enter fullscreen": "フルスクリーンに入る",
-  "Workflow Logs": "ワークフローログ",
   "Table Viewer Could Not Be Loaded. Check if the data is valid.": "テーブルビューアを読み込めませんでした。データが有効か確認してください。",
   "3D Viewer Could Not Be Loaded. Check if the data is valid.": "3Dビューアを読み込めませんでした。データが有効か確認してください。",
   "2D Viewer Could Not Be Loaded. Check if the data is valid.": "2Dビューアを読み込めませんでした。データが有効か確認してください。",
-  "Output data": "出力データ",
-  "Select Data to Preview": "プレビューするデータを選択",
-  "Table Viewer": "テーブルビューア",
   "2D Viewer": "2Dビューア",
   "3D Viewer": "3Dビューア",
   "Your workflow completed without any output data.": "ワークフローは出力データなしで完了しました。",
   "Please review the logs to see if there were any errors during the workflow process.": "ワークフロープロセス中にエラーが発生したかどうかを確認するには、ログを確認してください。",
   "OK": "OK",
+  "Output data": "出力データ",
+  "Debug Run": "デバッグ実行",
+  "Workflow Logs": "ワークフローログ",
+  "Data Preview": "データプレビュー",
+  "Tabular Viewer": "表形式ビューア",
+  "Select Data to Preview": "プレビューするデータを選択",
   "Edge ID: ": "エッジID: ",
   "Source Node ID: ": "ソースノードID: ",
   "Target Node ID: ": "ターゲットノードID: ",
