+++ ui/src/lib/i18n/locales/zh.json
@@ -72,18 +72,20 @@
   "All nodes in viewport": "视图中的所有节点",
   "Exit fullscreen": "退出全屏",
   "Enter fullscreen": "进入全屏",
-  "Workflow Logs": "工作流日志",
   "Table Viewer Could Not Be Loaded. Check if the data is valid.": "",
   "3D Viewer Could Not Be Loaded. Check if the data is valid.": "",
   "2D Viewer Could Not Be Loaded. Check if the data is valid.": "",
-  "Output data": "输出数据",
-  "Select Data to Preview": "选择要预览的数据",
-  "Table Viewer": "表格查看器",
   "2D Viewer": "",
   "3D Viewer": "3D查看器",
   "Your workflow completed without any output data.": "您的工作流完成，但没有任何输出数据。",
   "Please review the logs to see if there were any errors during the workflow process.": "请查看日志，以查看工作流过程中是否有任何错误。",
   "OK": "确定",
+  "Output data": "输出数据",
+  "Debug Run": "",
+  "Workflow Logs": "工作流日志",
+  "Data Preview": "",
+  "Tabular Viewer": "",
+  "Select Data to Preview": "选择要预览的数据",
   "Edge ID: ": "边缘ID: ",
   "Source Node ID: ": "源节点ID: ",
   "Target Node ID: ": "目标节点ID: ",
